requirements.txt
torch 
torchvision
albumentations
pytorch-lightning
timm
kaggle
modal-client
  